requirements.txt
psycopg2-binary==2.9.9
PyQt5==5.15.10
python-dotenv==1.0.0
requests==2.31.0
pyinstaller==6.2.0

pip freeze > requirements.txt
pip install psycopg2
pip install PyQt5
pip install pyinstaller
pip install python-dotenv
pip install requests
pip install PyQt5-sip

tk
sqlite3
hashlib
smtplib
import smtplib
import ssl
